Renamed package for consistency (#186)
